# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Debt Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  super_admin: "super_admin",
  owner: "owner",
} as const;

export interface User {
  id: number;
  nama: string;
  username: string;
  role: UserRole;
  usaha_id?: number | null;
  is_active: boolean;
  created_at: string;
}

export interface LoginResponse {
  user: User;
  message: string;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  super_admin: "super_admin",
  owner: "owner",
} as const;

export interface CreateUserBody {
  nama: string;
  username: string;
  password: string;
  role: CreateUserBodyRole;
  usaha_id?: number | null;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  super_admin: "super_admin",
  owner: "owner",
} as const;

export interface UpdateUserBody {
  nama?: string;
  username?: string;
  role?: UpdateUserBodyRole;
  usaha_id?: number | null;
  is_active?: boolean;
}

export interface ResetPasswordBody {
  new_password: string;
}

export interface Usaha {
  id: number;
  nama_usaha: string;
  alamat?: string | null;
  telepon?: string | null;
  catatan?: string | null;
  created_at: string;
}

export interface CreateUsahaBody {
  nama_usaha: string;
  alamat?: string | null;
  telepon?: string | null;
  catatan?: string | null;
}

export interface Pelanggan {
  id: number;
  usaha_id: number;
  nama: string;
  telepon?: string | null;
  alamat?: string | null;
  catatan?: string | null;
  created_at: string;
}

export type HutangStatus = (typeof HutangStatus)[keyof typeof HutangStatus];

export const HutangStatus = {
  aktif: "aktif",
  lunas: "lunas",
} as const;

export interface Hutang {
  id: number;
  usaha_id: number;
  pelanggan_id: number;
  pelanggan_nama: string;
  tanggal_hutang: string;
  keterangan?: string | null;
  nominal_hutang: number;
  total_dibayar: number;
  sisa_hutang: number;
  status: HutangStatus;
  created_at: string;
  updated_at: string;
}

export interface Pembayaran {
  id: number;
  usaha_id: number;
  hutang_id: number;
  pelanggan_id: number;
  pelanggan_nama: string;
  tanggal_bayar: string;
  nominal_bayar: number;
  catatan?: string | null;
  created_at: string;
}

export interface PelangganDetail {
  id: number;
  usaha_id: number;
  nama: string;
  telepon?: string | null;
  alamat?: string | null;
  catatan?: string | null;
  created_at: string;
  hutang_list: Hutang[];
  pembayaran_list: Pembayaran[];
  total_hutang: number;
  total_dibayar: number;
  sisa_hutang: number;
}

export interface CreatePelangganBody {
  nama: string;
  telepon?: string | null;
  alamat?: string | null;
  catatan?: string | null;
}

export type HutangDetailStatus =
  (typeof HutangDetailStatus)[keyof typeof HutangDetailStatus];

export const HutangDetailStatus = {
  aktif: "aktif",
  lunas: "lunas",
} as const;

export interface HutangDetail {
  id: number;
  usaha_id: number;
  pelanggan_id: number;
  pelanggan_nama: string;
  tanggal_hutang: string;
  keterangan?: string | null;
  nominal_hutang: number;
  total_dibayar: number;
  sisa_hutang: number;
  status: HutangDetailStatus;
  created_at: string;
  updated_at: string;
  pembayaran_list: Pembayaran[];
}

export interface CreateHutangBody {
  pelanggan_id: number;
  tanggal_hutang: string;
  keterangan?: string | null;
  nominal_hutang: number;
}

export interface UpdateHutangBody {
  tanggal_hutang?: string;
  keterangan?: string | null;
  nominal_hutang?: number;
}

export interface CreatePembayaranBody {
  hutang_id: number;
  tanggal_bayar: string;
  nominal_bayar: number;
  catatan?: string | null;
}

export interface OwnerDashboard {
  total_hutang: number;
  total_dibayar: number;
  sisa_hutang: number;
  jumlah_pelanggan_berhutang: number;
  jumlah_hutang_aktif: number;
  jumlah_hutang_lunas: number;
  pembayaran_terbaru: Pembayaran[];
  hutang_terbesar: Hutang[];
}

export interface UsahaHutangSummary {
  id: number;
  nama_usaha: string;
  sisa_hutang: number;
  jumlah_pelanggan: number;
}

export interface AdminDashboard {
  jumlah_usaha: number;
  jumlah_owner_aktif: number;
  total_pelanggan: number;
  total_hutang: number;
  total_dibayar: number;
  total_sisa_hutang: number;
  usaha_terbaru: Usaha[];
  usaha_hutang_terbesar: UsahaHutangSummary[];
}

export type LaporanItemStatus =
  (typeof LaporanItemStatus)[keyof typeof LaporanItemStatus];

export const LaporanItemStatus = {
  aktif: "aktif",
  lunas: "lunas",
} as const;

export interface LaporanItem {
  tanggal_hutang: string;
  nama_pelanggan: string;
  keterangan?: string | null;
  nominal_hutang: number;
  total_dibayar: number;
  sisa_hutang: number;
  status: LaporanItemStatus;
}

export interface BackupData {
  version: string;
  exported_at: string;
  usaha_id: number;
  usaha: Usaha;
  pelanggan: Pelanggan[];
  hutang: Hutang[];
  pembayaran: Pembayaran[];
}

export type GetHutangListParams = {
  pelanggan_id?: number;
  status?: GetHutangListStatus;
};

export type GetHutangListStatus =
  (typeof GetHutangListStatus)[keyof typeof GetHutangListStatus];

export const GetHutangListStatus = {
  aktif: "aktif",
  lunas: "lunas",
} as const;

export type GetPembayaranListParams = {
  hutang_id?: number;
  pelanggan_id?: number;
};

export type GetLaporanParams = {
  pelanggan_id?: number;
  status?: GetLaporanStatus;
  tanggal_dari?: string;
  tanggal_sampai?: string;
};

export type GetLaporanStatus =
  (typeof GetLaporanStatus)[keyof typeof GetLaporanStatus];

export const GetLaporanStatus = {
  aktif: "aktif",
  lunas: "lunas",
} as const;