# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Debt Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /hutang:
    get:
      operationId: getHutangList
      tags: [hutang]
      summary: Get all debts for owner's business
      parameters:
        - name: pelanggan_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [aktif, lunas]
      responses:
        "200":
          description: List of debts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hutang"
    post:
      operationId: createHutang
      tags: [hutang]
      summary: Create a debt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHutangBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hutang"
  /hutang/{id}:
    get:
      operationId: getHutang
      tags: [hutang]
      summary: Get a debt with payment history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Debt detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HutangDetail"
    put:
      operationId: updateHutang
      tags: [hutang]
      summary: Update a debt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHutangBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hutang"
    delete:
      operationId: deleteHutang
      tags: [hutang]
      summary: Delete a debt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /backup/restore:
    post:
      operationId: importBackup
      tags: [backup]
      summary: Restore owner data from JSON backup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackupData"
      responses:
        "200":
          description: Restore result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user, message]
    User:
      type: object
      properties:
        id:
          type: integer
        nama:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [super_admin, owner]
        usaha_id:
          type: integer
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
      required: [id, nama, username, role, is_active, created_at]
...
    Usaha:
      type: object
      properties:
        id:
          type: integer
        nama_usaha:
          type: string
        alamat:
          type: string
          nullable: true
        telepon:
          type: string
          nullable: true
        catatan:
          type: string
          nullable: true
        created_at:
          type: string
      required: [id, nama_usaha, created_at]
...
    Pelanggan:
      type: object
      properties:
        id:
          type: integer
        usaha_id:
          type: integer
        nama:
          type: string
        telepon:
          type: string
          nullable: true
        alamat:
          type: string
          nullable: true
        catatan:
          type: string
          nullable: true
        created_at:
          type: string
      required: [id, usaha_id, nama, created_at]
...
    Hutang:
      type: object
      properties:
        id:
          type: integer
        usaha_id:
          type: integer
        pelanggan_id:
          type: integer
        pelanggan_nama:
          type: string
        tanggal_hutang:
          type: string
        keterangan:
          type: string
          nullable: true
        nominal_hutang:
          type: number
        total_dibayar:
          type: number
        sisa_hutang:
          type: number
        status:
          type: string
          enum: [aktif, lunas]
        created_at:
          type: string
        updated_at:
          type: string
      required: [id, usaha_id, pelanggan_id, pelanggan_nama, tanggal_hutang, nominal_hutang, total_dibayar, sisa_hutang, status, created_at, updated_at]
    HutangDetail:
      type: object
      properties:
        id:
          type: integer
        usaha_id:
          type: integer
        pelanggan_id:
          type: integer
        pelanggan_nama:
          type: string
        tanggal_hutang:
          type: string
        keterangan:
          type: string
          nullable: true
        nominal_hutang:
          type: number
        total_dibayar:
          type: number
        sisa_hutang:
          type: number
        status:
          type: string
          enum: [aktif, lunas]
        created_at:
          type: string
        updated_at:
          type: string
        pembayaran_list:
          type: array
          items:
            $ref: "#/components/schemas/Pembayaran"
      required: [id, usaha_id, pelanggan_id, pelanggan_nama, tanggal_hutang, nominal_hutang, total_dibayar, sisa_hutang, status, created_at, updated_at, pembayaran_list]
    CreateHutangBody:
      type: object
      properties:
        pelanggan_id:
          type: integer
        tanggal_hutang:
          type: string
        keterangan:
          type: string
          nullable: true
        nominal_hutang:
          type: number
      required: [pelanggan_id, tanggal_hutang, nominal_hutang]
    UpdateHutangBody:
      type: object
      properties:
        tanggal_hutang:
          type: string
        keterangan:
          type: string
          nullable: true
        nominal_hutang:
          type: number
    Pembayaran:
      type: object
      properties:
        id:
          type: integer
        usaha_id:
          type: integer
        hutang_id:
          type: integer
        pelanggan_id:
          type: integer
        pelanggan_nama:
          type: string
        tanggal_bayar:
          type: string
        nominal_bayar:
          type: number
        catatan:
          type: string
          nullable: true
        created_at:
          type: string
      required: [id, usaha_id, hutang_id, pelanggan_id, pelanggan_nama, tanggal_bayar, nominal_bayar, created_at]
...
    BackupData:
      type: object
      properties:
        version:
          type: string
        exported_at:
          type: string
        usaha_id:
          type: integer
        usaha:
          $ref: "#/components/schemas/Usaha"
        pelanggan:
          type: array
          items:
            $ref: "#/components/schemas/Pelanggan"
        hutang:
          type: array
          items:
            $ref: "#/components/schemas/Hutang"
        pembayaran:
          type: array
          items:
            $ref: "#/components/schemas/Pembayaran"
      required: [version, exported_at, usaha_id, usaha, pelanggan, hutang, pembayaran]