# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Debt Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  nama: string;
  username: string;
  role: UserRole;
  usaha_id?: number | null;
  is_active: boolean;
  created_at: string;
}